投稿したユーザーのみ削除する機能を追加

+++ app/Http/Controllers/CommentController.php
@@ -27,6 +27,7 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        $this->authorize('delete', '$comment');
         $comment->delete();
         return back();
     }
